Move model loading out of renderer
