+++ .clang-format
@@ -1,2 +1,3 @@
 BasedOnStyle: Microsoft
 BreakBeforeBraces: Mozilla
+AllowShortFunctionsOnASingleLine: InlineOnly
